+++ src/DataSource/src/AbstractEntity.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace spaceonfire\DataSource;
 
-use spaceonfire\DataSource\Adapters\NetteUtils\SmartArrayAccessObject;
+use spaceonfire\DataSource\Bridge\NetteUtils\SmartArrayAccessObject;
 
 abstract class AbstractEntity implements EntityInterface
 {