+++ src/DataSource/src/Adapters/CycleOrm/Mapper/Hydrator/StdClassHydrator.php
@@ -4,53 +4,19 @@ declare(strict_types=1);
 
 namespace spaceonfire\DataSource\Adapters\CycleOrm\Mapper\Hydrator;
 
-use Laminas\Hydrator\AbstractHydrator;
+use function class_alias;
 
-class StdClassHydrator extends AbstractHydrator
-{
+class_alias(
+    \spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\StdClassHydrator::class,
+    __NAMESPACE__ . '\StdClassHydrator'
+);
+
+if (false) {
     /**
-     * @inheritDoc
+     * @deprecated Will be dropped in next major release.
+     * Use \spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\StdClassHydrator instead.
      */
-    public function extract(object $object): array
+    class StdClassHydrator extends \spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\StdClassHydrator
     {
-        $data = get_object_vars($object);
-        $filter = $this->getFilter();
-
-        /**
-         * @var string $name
-         * @var mixed $value
-         */
-        foreach ($data as $name => $value) {
-            // Filter keys, removing any we don't want
-            if (!$filter->filter($name)) {
-                unset($data[$name]);
-                continue;
-            }
-
-            // Replace name if extracted differ
-            $extracted = $this->extractName($name, $object);
-
-            if ($extracted !== $name) {
-                unset($data[$name]);
-                $name = $extracted;
-            }
-
-            $data[$name] = $this->extractValue($name, $value, $object);
-        }
-
-        return $data;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function hydrate(array $data, object $object)
-    {
-        foreach ($data as $name => $value) {
-            $property = $this->hydrateName($name, $data);
-            $object->$property = $this->hydrateValue($property, $value, $data);
-        }
-
-        return $object;
     }
 }