+++ src/DataSource/src/Adapters/CycleOrm/Mapper/Hydrator/Strategy/CollectionExtractorStrategy.php
@@ -4,56 +4,19 @@ declare(strict_types=1);
 
 namespace spaceonfire\DataSource\Adapters\CycleOrm\Mapper\Hydrator\Strategy;
 
-use Cycle\ORM\Promise\Collection\CollectionPromiseInterface;
-use Cycle\ORM\Promise\ReferenceInterface;
-use Cycle\ORM\Relation\HasMany;
-use Cycle\ORM\Relation\ManyToMany;
-use Cycle\ORM\Relation\Pivoted\PivotedCollection;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection as DoctrineCollection;
-use Laminas\Hydrator\Strategy\StrategyInterface;
-use spaceonfire\Collection\CollectionInterface;
-use spaceonfire\DataSource\Adapters\CycleOrm\Collection\PivotAwareInterface;
-use Webmozart\Assert\Assert;
+use function class_alias;
 
-class CollectionExtractorStrategy implements StrategyInterface
-{
+class_alias(
+    \spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\Strategy\CollectionExtractorStrategy::class,
+    __NAMESPACE__ . '\CollectionExtractorStrategy'
+);
+
+if (false) {
     /**
-     * @inheritDoc
-     *
-     * Convert spaceonfire collection to doctrine collection for proper relations handling by Cycle ORM
-     *
-     * @param CollectionInterface|null|mixed $value
-     * @return DoctrineCollection|mixed
-     * @see HasMany::extract()
-     * @see ManyToMany::extract()
-     * @see https://github.com/cycle/orm/issues/24
+     * @deprecated Will be dropped in next major release.
+     * Use \spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\Strategy\CollectionExtractorStrategy instead.
      */
-    public function extract($value, ?object $object = null)
+    class CollectionExtractorStrategy extends \spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\Strategy\CollectionExtractorStrategy
     {
-        if ($value instanceof ReferenceInterface) {
-            return $value;
-        }
-
-        if ($value instanceof CollectionPromiseInterface) {
-            return $value->getPromise();
-        }
-
-        Assert::nullOrIsInstanceOf($value, CollectionInterface::class);
-        $elements = $value ? $value->all() : [];
-
-        if ($value instanceof PivotAwareInterface) {
-            return new PivotedCollection($elements, $value->getPivotContext());
-        }
-
-        return new ArrayCollection($elements);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function hydrate($value, ?array $data)
-    {
-        return $value;
     }
 }