+++ src/DataSource/tests/Bridge/CycleOrm/Query/CycleQueryExpressionVisitorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace spaceonfire\DataSource\Adapters\CycleOrm\Query;
+namespace spaceonfire\DataSource\Bridge\CycleOrm\Query;
 
 use Cycle\ORM\Select\QueryBuilder;
 use Cycle\ORM\Select\RootLoader;
 use InvalidArgumentException;
 use spaceonfire\Criteria\Expression\AbstractExpressionDecorator;
-use spaceonfire\DataSource\Adapters\CycleOrm\AbstractCycleOrmTest;
+use spaceonfire\DataSource\Bridge\CycleOrm\AbstractCycleOrmTest;
 use spaceonfire\DataSource\Fixtures\Infrastructure\Mapper\StubMapper;
 use Spiral\Database\Query\SelectQuery;
 use Webmozart\Expression\Expr;