+++ src/DataSource/tests/Fixtures/Domain/Post/Post.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace spaceonfire\DataSource\Fixtures\Domain\Post;
 
-use spaceonfire\DataSource\Adapters\NetteUtils\SmartArrayAccessObject;
+use spaceonfire\DataSource\Bridge\NetteUtils\SmartArrayAccessObject;
 use spaceonfire\DataSource\EntityInterface;
 use spaceonfire\DataSource\JsonSerializableObjectTrait;